+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## dev
 
+* Skip empty blocks with `getContent` (#204)
 ## 0.12.3 - 2016-05-26
 
 * Add label to Author & Publisher dropdown (#203)